[front] refactor: display the extension section inside the reco. section
