+++ frontend/src/pages/home/videos/HomeVideos.tsx
@@ -10,8 +10,7 @@ import AlternatingBackgroundColorSectionList from 'src/pages/home/AlternatingBac
 import TitleSection from 'src/pages/home/TitleSection';
 import PollListSection from 'src/pages/home/PollListSection';
 import ComparisonSection from 'src/pages/home/videos/sections/ComparisonSection';
-import ExtensionSection from 'src/pages/home/videos/sections/ExtensionSection';
-import RecommendationsSection from 'src/pages/home/videos/sections/RecommendationsSection';
+import RecommendationsSection from 'src/pages/home/videos/sections/recommendations/RecommendationsSection';
 import ResearchSection from 'src/pages/home/videos/sections/research/ResearchSection';
 import { DEFAULT_POLL_STATS, getPollStats } from 'src/utils/api/stats';
 import { PollStats } from 'src/utils/types';
@@ -121,7 +120,6 @@ const HomeVideosPage = () => {
         }}
       />
       <ResearchSection />
-      <ExtensionSection />
       <UsageStatsSection externalData={stats} />
       <PollListSection />
     </AlternatingBackgroundColorSectionList>
